+++ internal/analyzer/analyze.go
@@ -141,7 +141,7 @@ func (ks *KScanner) checkKubernetesList(ctx context.Context) error {
 	log.Printf(config.Yellow("Begin Job and CronJob analyzing"))
 	log.Printf(config.Yellow("Begin DaemonSet analyzing"))
 
-	if ctx.Value("nameSpace") == "all" {
+	if ctx.Value("nameSpace") == "all" || ctx.Value("nameSpace") != "standard" {
 		namespaceWhileList = []string{}
 	}
 
@@ -191,22 +191,20 @@ func (ks *KScanner) checkKubernetesList(ctx context.Context) error {
 				}
 			}
 
+			err = ks.checkConfigMap(ns.Name)
+			if err != nil {
+				log.Printf("check config map failed in namespace: %s, %v", ns.Name, err)
+			}
+
+			err = ks.checkSecret(ns.Name)
+			if err != nil {
+				log.Printf("check secret failed in namespace %s, %v", ns.Name, err)
+			}
+
 			if isNecessary {
 				err = ks.checkRoleBinding(ns.Name)
 				if err != nil {
 					log.Printf("check role binding failed in namespace: %s, %v", ns.Name, err)
-				}
-
-				// TODO: remove from the white list, add kube-system namespace checking
-				err = ks.checkConfigMap(ns.Name)
-				if err != nil {
-					log.Printf("check config map failed in namespace: %s, %v", ns.Name, err)
-				}
-
-				// TODO: remove from the white list, add kube-system namespace checking
-				err = ks.checkSecret(ns.Name)
-				if err != nil {
-					log.Printf("check secret failed in namespace %s, %v", ns.Name, err)
 				}
 
 				err := ks.checkPod(ns.Name)