+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 1.0.6 (2023.3.2)
 ## features
 - Add Kubernetes `DaemonSet` checking
-- Add rootkit and backdoor checking
+- Add rootkit and backdoor checking in K8s and Docker
 
 ## improvements
 - Add some rules for CAP checking
+- Change the namespace checking of Secret and ConfigMap
+- Improve the rules of `DeamonSet` scanning
 
 # 1.0.5 (2023.2.13)
 ## features